+++ locale/easter.js
@@ -1,19 +1,54 @@
 //! moment-holiday.js locale configuration
-//! locale : Easter (Adds Good Friday and Easter Sunday to any moment-holiday locale)
+//! locale : Easter Related Holidays
 //! author : Kodie Grantham : https://github.com/kodie
 
 (function() {
   var moment = (typeof require !== 'undefined' && require !== null) && !require.amd ? require('moment') : this.moment;
 
   moment.holidays.easter = {
+    "Ash Wednesday": {
+      date: 'ashwednesday',
+      keywords_y: ['ash'],
+      keywords: ['wednesday']
+    },
+    "Lent": {
+      date: 'lent',
+      keywords_y: ['lent']
+    },
     "Good Friday": {
       date: 'goodfriday',
       keywords_y: ['good', 'friday']
+    },
+    "Holy Saturday": {
+      date: 'holysaturday',
+      keywords_y: ['holy', 'saturday']
     },
     "Easter Sunday": {
       date: 'easter',
       keywords_y: ['easter'],
       keywords: ['sunday']
+    },
+    "Easter Monday": {
+      date: 'eastermonday',
+      keywords_y: ['easter', 'monday']
+    },
+    "Ascension Day": {
+      date: 'ascension',
+      keywords_y: ['ascension']
+    },
+    "Pentecost Sunday": {
+      date: 'pentecost',
+      keywords_y: ['pentecost'],
+      keywords: ['sunday']
+    },
+    "Whit Monday": {
+      date: 'whitmonday',
+      keywords_y: ['whit'],
+      keywords: ['monday']
+    },
+    "Corpus Christi": {
+      date: 'corpuschristi',
+      keywords: ['corpus', 'christi', 'feast']
     }
   };
 
@@ -33,10 +68,41 @@
   };
 
   moment.modifyHolidays.extendParser(function(m, date){
-    if (date === 'easter' || date === 'goodfriday') {
+    if (date === 'ashwednesday' ||
+        date === 'lent' ||
+        date === 'goodfriday' ||
+        date === 'holysaturday' ||
+        date === 'easter' ||
+        date === 'eastermonday' ||
+        date === 'ascension' ||
+        date === 'pentecost' ||
+        date === 'whitmonday' ||
+        date === 'corpuschristi') {
+
       var e = easter(m.year());
-      if (date === 'easter') { return e; }
+
+      if (date === 'ashwednesday') { return e.subtract(46, 'days'); }
       if (date === 'goodfriday') { return e.subtract(2, 'days'); }
+      if (date === 'holysaturday') { return e.subtract(1, 'day'); }
+      if (date === 'easter') { return e; }
+      if (date === 'eastermonday') { return e.add(1, 'day'); }
+      if (date === 'ascension') { return e.add(39, 'days'); }
+      if (date === 'pentecost') { return e.add(49, 'days'); }
+      if (date === 'whitmonday') { return e.add(50, 'days'); }
+      if (date === 'corpuschristi') { return e.add(60, 'days'); }
+
+      if (date === 'lent') {
+        var lent = [];
+        var d = moment(e).subtract(46, 'days');
+        var length = e.subtract(2, 'days').diff(d, 'days');
+
+        for (i = 0; i < length; i++) {
+          lent.push(moment(d));
+          d.add(1, 'day');
+        }
+
+        return lent;
+      }
     }
   });